pyproject.toml
[project]
name = "col851-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chronos-forecasting>=1.5.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "prometheus-client>=0.23.1",
    "scikit-learn>=1.7.2",
    "torch>=2.8.0",
]
